utils/requirements.txt
torch>=2.0.0
transformers>=4.30.0
supervision>=0.20.0
albumentations>=1.3.0
pycocotools>=2.0.6
torchmetrics>=1.0.0
matplotlib>=3.5.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.65.0
opencv-python>=4.7.0

# SPDNet dependencies for de-raining preprocessing
pytorch_wavelets>=1.3.0
PyWavelets>=1.0.0
